refactor(storage-manager): store sub-intervals in BTreeMap
This commit changes the data structure used to store the `SubInterval`s
within an `Interval` from a `HashMap` to a `BTreeMap`.

The main purpose of this change is to diminish the number of operations
required to search for events: given indexes for `Interval` and
`SubInterval`, a wildcard update or delete need to be applied
"backward", i.e. on any event that belongs in an `Interval` of lower or
equal index (and in the `Interval` of equal index, in any `SubInterval`
with a lower or equal index). As a `BTreeMap` stores its entries in
order, this allows us to stop considering `Interval` / `SubInterval` the
moment the index is greater than that of the considered event.

* plugins/zenoh-plugin-storage-manager/src/replication/classification.rs:
  changed, in the `Interval` structure, the data structure used to store
  the `sub_intervals` to a `BTreeMap`.

Signed-off-by: Julien Loudet <julien.loudet@zettascale.tech>