refactor(storage-manager): change signature of the `prefix` function
This commit changes the signature of the `prefix` function to accept
`Option` for both the prefix and the key expression to
prefix. Conversely, this function now returns an error if both are set
to `None`.

This change is in anticipation of the support for Wildcard updates when
replicating Storage: in order to check if a Wildcard should be applied
on a key expression, we need to `prefix` the key. This change avoids
repeating the checks and raising this error at all the call sites.

* plugins/zenoh-plugin-storage-manager/src/lib.rs: changed the signature
  of the `prefix` function to accept two `Option` and return a `Result`.
* plugins/zenoh-plugin-storage-manager/src/storages_mgt/service.rs:
  updated the call to `prefix`.

Signed-off-by: Julien Loudet <julien.loudet@zettascale.tech>